Adding an Accessibility tab alert for upgraders
See GitHub issue #14; upgrading to the latest beta has left Jumpcut in a
state for some testers where the program is listed in the Accessibility
security tab as authorized, but the paste functionality (which
Accessibility authorizes) is still not working. Removing and re-adding
has fixed the issue, seemingly for good. (This problem seems to be
specific to people *upgrading*.) We attempt to be helpful by adding a
one-time alert message to upgraders, informing that they should try this
manually if the paste behavior is broken.