Preventing bezel from triggering when setting hotkey
We use a recipe from the ShortcutRecorder2 project to disable listening
for hotkey events while the user is setting a hotkey. This prevents the
edge case of the bezel triggering if users set the same hotkey that
they've already used. (Unlikely to come up that often during normal use,
but something that happens fairly frequently in testing and which is
always annoying.)