Mojave (and Dark Mode) support in ShortcutRecorder control
Jumpcut is using an older release of ShortcutRecorder2; this change
grabs Mojave-specific assets and makes some changes internally to bring
support for Dark Mode within the hotkey widget. Most of the relevant
work was done within the snark/ShortcutRecorder2 fork on GitHub, but
all the assets and hard work are due to the ShortcutRecorder2 team.