Push required version of macOS back to 10.11 (Yosemite)
Previous builds were using High Sierra as the required version of macOS
to run the 0.7x branch of Jumpcut, just because that's what I'm
developing on. 10.11 Yosemite seems to be as far as I can push this back
before Xcode starts giving me version-specific warnings, so we'll push
it back that far.