Github issue #30: Crash for Japanese keyboards
keyCodeToString, used when finding the key code for "v" (which in
turn is used to issue the command-v key calls to other applications)
relied on the TISCopyCurrentKeyboardInputSource method; this returns
a null for at least some non-Roman, causing a crash when switching
between Roman-style and Japanese-style keyboards.

Per a piece of MIT-licensed Microsoft code[1], there's an alternative
call available, TISCopyCurrentKeyboardLayoutInputSource. We use this
if the first call doesn't get a result, in addition to making the
entire function more defensive. (This is very much not a hot loop;
it runs on the application launch and then when the keyboard layout is
switched by the user.)

[1] https://github.com/microsoft/node-native-keymap/blob/master/src/keyboard_mac.mm