Switch to mainline ShortcutRecorder
In order to preserve older code using PTHotkey, Jumpcut was using a fork
of ShortcutRecorder based off ShortcutRecorder circa 2017. This worked
well for recent versions of macOS (post-High Sierra), but caused issues
on earlier versions, which are still officially supported. This
changeset moves Jumpcut to use a fork based on a more recent version
of ShortcutRecorder from October 2019, which should help with older
versions of macOS.