Improved detection of Jumpcut-created clips
This changeset updates the mechanism used for deciding which clippings
should not be recorded, resolving Github issue #48. Previously, an exact
comparison was done with the first clipping on the stack, meaning that
if clippings were pasted from further down in a way that *did not*
modify the stack, they were erroneously added back to the top of the
stack. We now set an internal marker on items added to the pasteboard
and test for the existence of this internal marker.

As part of this change, the checks in the pollPasteboard function were
refactored. Additionally, when adding to the stack, we now set the
org.nspasteboard.source marker to the null string, per the intended
behavior described at NSPasteboard.org. (We do not track the origin of
clippings, so we cannot set the source to the originating application.)
Finally, we are no longer ignoring all items with
a org.nspasteboard.source value; it's possible that this will lead to
unwanted and confusing interactions with other pasteboard-manipulating
actions, in which case that aspect of this changeset will be reverted.