Revised checks for "Concealed" and "Transient" pasteboard items
For unclear reasons, the detection method used to test for "transient"
and, optionally, "concealed" pasteboard item types was failing in some
cases. This change switches the detection method from an attempted load
of the item for a given type to direct comparison of the types, which
seems less brittle. (See GitHub issues nos. 10 and 55.)