Complete replacement of Objective-C with Swift code
This changeset represents a complete rewrite of Jumpcut in Swift 5. For
a variety of reasons, maintaining the Objective-C code for Jumpcut was
becoming increasingly difficult; this rewrite hopefully provides a clean
base for future work for myself and for any users who wish to run their
own forks! We no longer use internal forks of any libraries, and
everything has been switched over to Swift Package Manager from
vendoring. As part of the change, we bump the minimum OS target from
10.10 Yosemite to 10.11 El Capitan.

Along with the rewrite, we add two long-requested features. First, we
add a "headless" mode (Github issue #20), with no status icon in the
menu bar; when in headless mode, doubleclicking the application icon
will open the preferences window where headless mode may be disabled.
Second, we add a "toggle" behavior, where the paste/copy behavior is
overridden based on a modifier when clicking the status icon (GitHub
issue #33).

Along with these two features, we now offer an alternate "additional
options" menu, configurable to a modifier on menu click; this enables
users who prefer the menu to delete individual items and will likely be
the basis for future additional controls.

A "reset preferences" control is now provided, along with controls for
left- (or right-) aligning bezel text.

Finally, new language, new icon; the charming but amateurish Jumpcut
icon I made years and years ago is no longer in keeping with macOS
design trends, and looked quite bad at smaller sizes. We replace it with
something better-suited for this decade and post-Big Sur design
standards, created for Jumpcut by John Kenzie.