Restore "white scissors" and "solid scissors" menu icon options
We had switched to allowing only the default Jumpcut icon, well-loved
for this decade-plus, as the menu bar icon, but I got a request to
restore the "chunkier white scissors" for extra visibility. It's
essentially harmless to restore, so here we go. (The "black scissors"
emoji seems visually too tall for the menu icon, so we'll omit it at
this time.)

Resolves GitHub issue #68.