Bezel behavior regression fixed
As part of the 0.80 rewrite, a regression was introduced where new
clippings pushed the stack down but the bezel position was not adjusted.
This meant that users could *neither* immediately return to the
last-used clipping *nor* immediately see the most recent new copy when
using the bezel. Obviously this makes Jumpcut less useful. We now
default to the Jumpcut 0.75 behavior of the bezel showing the most
recently-added clipping, and expose via preference an option (which has
been available in Jumpcut's past, although not in 0.75) to instead stick
with the most recently-used clipping even after copying new items.

Separately, this changeset corrects an issue where the "rememberNum"
preference was being ignored. We now make use of this value when sizing
the clipping stack. (Unlike Jumpcut 0.75, this value is checked only at
startup, which is I believe acceptable.)