Restore number-key behavior for bezel
A feature left out of the 0.80 for lack of time (and vague thoughts of
making things more user-configurable) except for the use of "1" to
return to the top of the stack. Everyone correctly found this annoying!
We return the use of 1-9, with 0 representing 10; while we're at it, we
return support for home/end (moving to the top and bottom of the stack)
and page up/page down (moving up and down 10 clippings at a time).