"skipSave" preference
This commit restores an option to run Jumpcut's stack wholly in-memory,
skipping reading or writing to the disk. (This differs from the pre-0.80
options, which included an option to save only on exit; that was
something that made sense in the days of spinning metal hard drives, but
doesn't make sense now.)

For both UX--what if someone changes in the middle of a session? what if
someone loses their data by misclicking?--this is currently *not*
exposed in the Preferences window; instead it requires use of the
`defaults` command-line application to set skipSafe to the value of
`true`. We may surface this value in the future.

GitHub issue #73.