Special-cased escape key handler for the bezel
Because the default behavior of NSResponder binds the escape key to
cancelOperation, we don't automatically pass through "keyDown" when
hitting escape. (For unresolved reasons, this seems to occur only
when control is also depressed, but that's the underlying cause.)
This changeset adds a performKeyEquivalent override that special-cases
escape, allowing us to use our regular keyDown handling to handle bezel
dismissal.

GitHub issue #77.