Fix for dual-monitor bezel display issues
This corrects two problems with using Jumpcut with multiple monitors
enabled, GitHub issues #54 and #76. The first was that Jumpcut always
displayed the bezel on the first screen it came up on; this has been
resolved by always setting the bezel to display on the main (active)
monitor. The second is more obscure; when launching Jumpcut with two
monitors active, using the hotkey did not actually display the bezel;
the underlying cause of this issue is still unclear but it's been
confirmed resolved by starting on screen zero on the very first use of
the bezel, then switching to mainScreen per usual.

Thanks to Salmeen Majid for reporting and tetsing the fix for the second
issue!