Full exposure of the skipSave preference
We expose a checkbox for skipSave, as though it were a normal preference
in the Clippings preference pane--because of weird edge cases, however,
we require the user to clear the clippings before toggling the setting
(and additionally provide a confirmation modal when disabling saving).

In order to provide this, we create a new control, "FixedButton", which
overrides the normal click behavior; it changes its state only
programmatically.

Additionally, we shunt some longer text into a new Constants file.