Remove headless check on double-click
There's currently no good support from Apple for preventing status bar
items from drifting into the notch from the right side; until such
a time as Jumpcut can produce an alert and let people know what's
happening, this means that a user can end up in a scenario where they
are unable to set the hotkey, etc., to use the application.

This change removes the check for headless mode when double-clicking
the application item, so that the Preferences window is always
accessible, even if the status bar item is not. (More thought is needed
about how to deal with the status bar icon/notch interactions.)