Switch out handrolled launch-on-login code
Maintaining a working launch-on-login environment that handles behavior
correctly for users across a wide variety of macOS versions is becoming
increasingly difficult given the amount of time available to spend on
Jumpcut bugfixes. As such, this changeset introduces use of Sindre
Sorhus' LaunchOnLogin library, which has resolved many of these issues.

Notably, this means that macOS 10.12 (Sierra) is now the oldest version
of macOS/OS X that can run Jumpcut. Users on El Capitan or earlier are
encouraged to remain on Jumpcut 0.82. We use a slightly older version of
LaunchOnLogin, to avoid cutting off a wider range of older Macintoshes.