Option to allow whitespace clippings
This feature was requested as GitHub issue #71, with a couple different
use cases noted (replacing single linebreaks with double linebreaks;
visual distinction within the menu). This changeset adds support via
a new "Allow Whitespace" option in the Clippings preference pane. I have
concerns about making the differences legible in the bezel and the menu
bar, but per my comment on the issue, they can be resolved at a future
date and this unblocks those use cases.