Updated file save behavior
We no longer use tilde string expansion; instead we use more modern
FileManager methods to find the Application Support directory.
Additionally, on launch and when setting the preference to save
clippings history (which is on by default), we show a warning alert if
Jumpcut does not have write access to the save final.

Finally, we do some light shuffling to consolidate a number of our alert
messages to a new Alerts file.